add more debug

+++ src/kafka/records.rs
@@ -60,7 +60,7 @@ pub struct KafkaRecord {
     pub length: i8,
     pub attributes: i8,
     pub timestamp_delta: i8,
-    pub offset_delta: i32,
+    pub offset_delta: i8,
     pub key: Vec<u8>,
     pub value: Vec<u8>,
     pub headers: Vec<KafkaRecordHeader>,
